+++ config/services/pay_by_link_payment/context_provider.php
@@ -4,22 +4,14 @@ declare(strict_types=1);
 
 namespace Symfony\Component\DependencyInjection\Loader\Configurator;
 
-use CommerceWeavers\SyliusTpayPlugin\ContextProvider\RegulationsUrlContextProvider;
 use CommerceWeavers\SyliusTpayPlugin\PayByLinkPayment\ContextProvider\BankListContextProvider;
 
 return static function(ContainerConfigurator $container): void {
     $services = $container->services();
 
-    $services->set(BankListContextProvider::class)
+    $services->set('commerce_weavers_sylius_tpay.pay_by_link_payment.context_provider.bank_list', BankListContextProvider::class)
         ->args([
             service('commerce_weavers_sylius_tpay.tpay.provider.validated_tpay_api_bank_list'),
-        ])
-        ->tag('sylius.ui.template_event.context_provider')
-    ;
-
-    $services->set(RegulationsUrlContextProvider::class)
-        ->args([
-            service('sylius.context.locale'),
         ])
         ->tag('sylius.ui.template_event.context_provider')
     ;
