+++ config/services/pay_by_link_payment/payum.php
@@ -1,0 +1,9 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Symfony\Component\DependencyInjection\Loader\Configurator;
+
+return static function(ContainerConfigurator $container): void {
+    $container->import('payum/**/*.php');
+};
