+++ config/services/validator.php
@@ -4,9 +4,9 @@ declare(strict_types=1);
 
 namespace Symfony\Component\DependencyInjection\Loader\Configurator;
 
+use CommerceWeavers\SyliusTpayPlugin\PayByLinkPayment\Validator\Constraint\ValidTpayChannelValidator;
 use CommerceWeavers\SyliusTpayPlugin\Validator\Constraint\EncodedGooglePayTokenValidator;
 use CommerceWeavers\SyliusTpayPlugin\Validator\Constraint\ForAuthorizedUserOnlyValidator;
-use CommerceWeavers\SyliusTpayPlugin\Validator\Constraint\ValidTpayChannelValidator;
 
 return static function(ContainerConfigurator $container): void {
     $services = $container->services();
