+++ src/PayByLinkPayment/Payum/Exception/UnableToGetBankListException.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace CommerceWeavers\SyliusTpayPlugin\Payum\Exception;
+namespace CommerceWeavers\SyliusTpayPlugin\PayByLinkPayment\Payum\Exception;
 
 use Tpay\OpenApi\Utilities\TpayException;
