+++ src/PayByLinkPayment/Payum/Request/GetTpayTransactionsChannels.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace CommerceWeavers\SyliusTpayPlugin\Payum\Request\Api;
+namespace CommerceWeavers\SyliusTpayPlugin\PayByLinkPayment\Payum\Request;
 
 use Payum\Core\Request\Generic;
