+++ src/PayByLinkPayment/Validator/Constraint/ValidTpayChannel.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace CommerceWeavers\SyliusTpayPlugin\Validator\Constraint;
+namespace CommerceWeavers\SyliusTpayPlugin\PayByLinkPayment\Validator\Constraint;
 
 use Symfony\Component\Validator\Constraint;
