+++ src/PayByLinkPayment/Validator/Constraint/ValidTpayChannelValidator.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace CommerceWeavers\SyliusTpayPlugin\Validator\Constraint;
+namespace CommerceWeavers\SyliusTpayPlugin\PayByLinkPayment\Validator\Constraint;
 
 use CommerceWeavers\SyliusTpayPlugin\Tpay\Provider\ValidTpayChannelListProviderInterface;
 use Symfony\Component\Validator\Constraint;
